Decompile func_func_80020344

+++ include/drawutil.h
@@ -19,7 +19,7 @@ void func_8001FF44(); // 8001ff44 fDrawScreenBorder
 void func_800200A0(int, char, char, char); // 800200a0 fDrawFullScreenFade
 void func_80020168();
 void func_800202DC(const char*, int, short, int); // 800202dc fDrawStringCentered
-void func_80020344(const char*, int, short, int); // short may be an int
+void func_80020344(const char*, int, int, int); // 80020344 fDrawStringCenteredBoxed
 void func_800203C4(const char*, int, short, int); // 800203c4 fDrawStringRightAligned
 void func_80020428(int, int, short, int); // 80020428 FUN_80020428
 void func_80020530(char*); // 80020530 FUN_80020530
